make sure test.sh's $tmpdir is set before checking it's a folder

+++ test.sh
@@ -7,7 +7,7 @@ trap cleanup EXIT
 
 cleanup() {
     pkill -P $$
-    if [ -d "$tmpdir" ]; then
+    if [ -n "$tmpdir" ] && [ -d "$tmpdir" ]; then
         rm -r "$tmpdir"
     fi
 }
